Extracted plan mapping to a separate component
